refactor: Move element parsers to reference directory - complete separation
Updated parser reorganization to properly separate concerns:

**Final Structure:**

src/txxt/parser/
├── reference/               ← Reference parser (self-contained)
│   ├── mod.rs              - Module declarations and re-exports
│   ├── parser.rs           - Main parser orchestration
│   ├── combinators.rs      - Parser combinators and primitives
│   ├── api.rs              - Public API entry points
│   ├── annotations.rs      - Annotation element parser
│   ├── definitions.rs      - Definition element parser
│   ├── document.rs         - Document element parser
│   ├── foreign.rs          - Foreign block element parser
│   ├── labels.rs           - Label element parser
│   ├── lists.rs            - List element parser
│   ├── parameters.rs       - Parameter element parser
│   └── sessions.rs         - Session element parser
│
├── grammarengine/          ← Grammar engine parser (self-contained)
│   ├── mod.rs
│   ├── engine.rs           - Tree walker and pattern orchestration
│   ├── regex_grammar_engine.rs - Generic regex pattern matching
│   ├── txxt_grammar.rs     - Grammar patterns for txxt language
│   └── unwrapper.rs        - Convert matched patterns to AST nodes
│
└── parser.rs               ← Top-level module with public API

**Key Changes:**

1. Moved element parsers from ast_build/ to reference/
2. Deleted redundant elements.rs re-export module
3. Updated all imports to use relative paths (super::)
4. Both parsers are now completely independent
5. No shared code between parser implementations

**Why This Structure:**

- Reference parser is self-contained with all its parsing logic
- Grammar engine parser is self-contained with its own unwrapper
- Easy to compile either parser independently
- Clear what code belongs to which implementation
- No coupling or circular dependencies between parsers
- Can choose/switch between implementations without affecting other

All 348+ tests passing with zero regressions.
