+++ src/txxt/parser/reference/definitions.rs
@@ -10,7 +10,7 @@ use std::sync::Arc;
 
 use crate::txxt::ast::{ContentItem, Definition, TextContent};
 use crate::txxt::lexer::Token;
-use crate::txxt::parser::reference::combinators::{
+use super::combinators::{
     aggregate_locations, byte_range_to_location, extract_tokens_to_text_and_location, token,
 };
