+++ src/txxt/parser/reference/lists.rs
@@ -10,7 +10,7 @@ use std::sync::Arc;
 
 use crate::txxt::ast::{ContentItem, List, ListItem, Location, TextContent};
 use crate::txxt::lexer::Token;
-use crate::txxt::parser::reference::combinators::{
+use super::combinators::{
     aggregate_locations, byte_range_to_location, compute_location_from_locations,
     extract_tokens_to_text_and_location, is_text_token, token,
 };
